fix(refactor): make CoderFromUrn methods private
